+++ py-polars/src/series.rs
@@ -385,11 +385,9 @@ impl PySeries {
         Ok(df.into())
     }
 
-    pub fn arg_unique(&self) -> PyResult<Py<PyArray1<u32>>> {
-        let gil = pyo3::Python::acquire_gil();
+    pub fn arg_unique(&self) -> PyResult<Self> {
         let arg_unique = self.series.arg_unique().map_err(PyPolarsEr::from)?;
-        let pyarray = PyArray1::from_vec(gil.python(), arg_unique);
-        Ok(pyarray.to_owned())
+        Ok(arg_unique.into_series().into())
     }
 
     pub fn arg_min(&self) -> Option<usize> {